Remove NO_PYTHON in ir.h
